requirements.txt
Django
Pillow
psycopg2-binary
djangorestframework
django-filter
drf-yasg